working ejs page and first round refactoring done
