+++ application/forecast/views/forecast.ejs
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Sneezl - Forecast</title>
+    <link rel="icon" type="image/x-icon" href="/assets/favicon.ico">
+    <link rel="shortcut icon" type="image/x-icon" href="/assets/favicon.ico">
+    <!--Link directly to the jQuery CDN-->
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+
+    <!--Import Google Icon Font-->
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+
+    <!--Import materialize.css-->
+    <link type="text/css" rel="stylesheet" href="css/materialize.min.css" media="screen,projection" />
+    <!-- Import custom material colours -->
+    <link type="text/css" rel="stylesheet" href="css/material-styles.css" media="screen,projection" />
+    <!-- Import custom styles for this page only -->
+    <link type="text/css" rel="stylesheet" href="css/forecast.css" media="screen,projection" />
+
+    <!--Let browser know website is optimized for mobile-->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
+    <style>
+        body {
+            display: flex;
+            min-height: 100vh;
+            flex-direction: column;
+        }
+
+        main {
+            flex: 1 0 auto;
+            padding: 20px;
+        }
+    </style>
+
+</head>
+
+<body style="background-color: var(--md-sys-color-surface-container-lowest);">
+
+    <%- include('../partials/navbar') %>
+
+        <div class="container">
+            <h3 class="center-align">5-Day Pollen Forecast</h3>
+            <div class="input-field">
+                <input id="latitude" type="text" placeholder="Latitude">
+                <input id="longitude" type="text" placeholder="Longitude">
+            </div>
+            <button id="getForecast" class="btn waves-effect waves-light">Get Forecast</button>
+            <p id="errorMessage" class="red-text"></p>
+            <div id="forecastContainer" style="margin-top: 20px;"></div>
+        </div>
+
+        <!--JavaScript at end of body for optimized loading-->
+        <script type="text/javascript" src="js/materialize.min.js"></script>
+        <script src="/js/forecast.js"></script>
+
+</body>
+
+</html>
