+++ application/partials/navbar.ejs
@@ -18,7 +18,7 @@
         </li>
         <li>
           <div>
-            <a href="forecast.html" style="color: var(--md-sys-color-on-secondary-container);">
+            <a href="/forecast" style="color: var(--md-sys-color-on-secondary-container);">
               <i class="material-icons left"
                 style="border-radius: 100px; color: var(--md-sys-color-on-secondary-container); background-color: var(--md-sys-color-secondary-container); padding-left: 15px; padding-right: 15px;">query_stats</i>
               <b>Forecast</b>
