--- application/pollenforecast/pollenapi.env
@@ -1,1 +1,0 @@
-GOOGLE_API_KEY=AIzaSyANWNr_rZKHj8hqKt8UP-M6hEIB_rw6zl4
