+++ package.json
@@ -24,7 +24,9 @@
     "nodemailer": "^6.9.16",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0",
-    "socket.io": "^4.8.1"
+    "socket.io": "^4.8.1",
+    "axios": "^1.7.9",
+    "cors": "^2.8.5"
   },
   "devDependencies": {
     "@types/ejs": "^3.1.5",
