--- public/forecast.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Forecast</title>
-  <link rel="icon" type="image/x-icon" href="/assets/favicon.ico">
-  <link rel="shortcut icon" type="image/x-icon" href="/assets/favicon.ico">
-  <!--Link directly to the jQuery CDN-->
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
-
-  <!--Import Google Icon Font-->
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-
-  <!--Import materialize.css-->
-  <link type="text/css" rel="stylesheet" href="css/materialize.min.css" media="screen,projection" />
-
-  <!-- Import custom material colours -->
-  <link type="text/css" rel="stylesheet" href="css/material-styles.css" media="screen,projection" />
-
-  <!--Let browser know website is optimized for mobile-->
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-</head>
-<body>
-
-  <!-- Dropdown Structure -->
-  <ul id="dropdown1" class="dropdown-content">
-    <li><a href="/my_account"><i class="material-icons left">person</i>Account</a></li>
-    <li class="divider"></li>
-    <li><a href="/diary"><i class="material-icons left">bookmarks</i>Diary</a></li>
-  </ul>
-  <nav class="teal lighten-2">
-    <div class="container">
-      <div class="nav-wrapper">
-        <a href="/"><img src="assets/app_logo.png" height="65"></a>
-        <ul class="right hide-on-med-and-down">
-          <li><a href="/" class="black-text"><i class="material-icons left black-text">travel_explore</i>Heatmap</a></li>
-          <li><a href="forecast.html" class="black-text"><i class="material-icons left black-text">query_stats</i>Forecast</a></li>
-          <li><a class="dropdown-trigger" href="#!" data-target="dropdown1"><i class="material-icons black-text">more_vert</i>
-            <i class="material-icons right">arrow_drop_down</i></a></li>
-        </ul>
-      </div>
-  </nav>
-
-  <div class="row center-align">
-    <H1>Forecast</H1>
-  </div>
-
-  <!--JavaScript at end of body for optimized loading-->
-  <script type="text/javascript" src="js/materialize.min.js"></script>
-  <script src="js/script.js"></script>
-
-</body>
-</html>
